+++ demo/flappy.js
@@ -68,6 +68,7 @@ scene("game", () => {
 			outline(4),
 			area(),
 			move(LEFT, SPEED),
+			cleanup(),
 			// give it tags to easier define behaviors see below
 			"pipe",
 		]);
@@ -79,6 +80,7 @@ scene("game", () => {
 			outline(4),
 			area(),
 			move(LEFT, SPEED),
+			cleanup(),
 			// give it tags to easier define behaviors see below
 			"pipe",
 			// raw obj just assigns every field to the game obj