+++ demo/shooter.js
@@ -212,6 +212,7 @@ scene("battle", () => {
 			color(127, 127, 255),
 			outline(4),
 			move(UP, BULLET_SPEED),
+			cleanup(),
 			// strings here means a tag
 			"bullet",
 		]);