+++ src/types.ts
@@ -412,6 +412,20 @@ interface KaboomCtx {
 	 */
 	move(direction: number | Vec2, speed: number): MoveComp,
 	/**
+	 * destroy() the character if it's out of screen. Optionally specify the amount of time it has to be off-screen before removal.
+	 *
+	 * @example
+	 * ```js
+	 * // remove this 3 seconds after it leaves screen
+	 * add([
+	 *     pos(80, 80),
+	 *     move(LEFT, 120),
+	 *     cleanup(3),
+	 * ]);
+	 * ```
+	 */
+	cleanup(time?: number): CleanupComp,
+	/**
 	 * Follow another game obj's position.
 	 */
 	follow(obj: Character | null, offset?: Vec2): FollowComp,
@@ -495,7 +509,7 @@ interface KaboomCtx {
 	 * ]);
 	 * ```
 	 */
-	lifespan(time: number, action?: () => void): LifespanComp,
+	lifespan(time: number, conf?: LifespanCompConf): LifespanComp,
 	/**
 	 * Register an event on all game objs with certain tag.
 	 *
@@ -2112,7 +2126,6 @@ interface Comp {
 type CharacterID = number;
 
 interface PosComp extends Comp {
-//  	id: "pos",
 	/**
 	 * Object's current world position.
 	 */
@@ -2139,7 +2152,6 @@ interface PosComp extends Comp {
 }
 
 interface ScaleComp extends Comp {
-//  	id: "scale",
 	scale: Vec2,
 	scaleTo(s: number): void,
 	scaleTo(s: Vec2): void,
@@ -2147,7 +2159,6 @@ interface ScaleComp extends Comp {
 }
 
 interface RotateComp extends Comp {
-//  	id: "rotate",
 	/**
 	 * Angle in degrees.
 	 */
@@ -2155,17 +2166,14 @@ interface RotateComp extends Comp {
 }
 
 interface ColorComp extends Comp {
-//  	id: "color",
 	color: Color,
 }
 
 interface OpacityComp extends Comp {
-//  	id: "opacity",
 	opacity: number,
 }
 
 interface OriginComp extends Comp {
-//  	id: "origin",
 	/**
 	 * Origin point for render.
 	 */
@@ -2173,7 +2181,6 @@ interface OriginComp extends Comp {
 }
 
 interface LayerComp extends Comp {
-//  	id: "layer",
 	/**
 	 * Which layer this game obj belongs to.
 	 */
@@ -2181,7 +2188,6 @@ interface LayerComp extends Comp {
 }
 
 interface ZComp extends Comp {
-//  	id: "z",
 	/**
 	 * Defines the z-index of this game obj
 	 */
@@ -2189,7 +2195,6 @@ interface ZComp extends Comp {
 }
 
 interface FollowComp extends Comp {
-//  	id: "follow",
 	follow: {
 		obj: Character,
 		offset: Vec2,
@@ -2197,7 +2202,9 @@ interface FollowComp extends Comp {
 }
 
 interface MoveComp extends Comp {
-//  	id: "move",
+}
+
+interface CleanupComp extends Comp {
 }
 
 type RectSide =
@@ -2231,7 +2238,6 @@ interface AreaCompConf {
 }
 
 interface AreaComp extends Comp {
-//  	id: "area",
 	/**
 	 * Collider area info.
 	 */
@@ -2342,7 +2348,6 @@ interface SpriteCompConf {
 }
 
 interface SpriteComp extends Comp {
-//  	id: "sprite",
 	/**
 	 * Width for sprite.
 	 */
@@ -2390,7 +2395,6 @@ interface SpriteComp extends Comp {
 }
 
 interface TextComp extends Comp {
-//  	id: "text",
 	/**
 	 * The text to render.
 	 */
@@ -2429,7 +2433,6 @@ interface TextCompConf {
 }
 
 interface RectComp extends Comp {
-//  	id: "rect",
 	/**
 	 * Width of rect.
 	 */
@@ -2449,7 +2452,6 @@ type AreaType =
 	;
 
 interface OutlineComp extends Comp {
-//  	id: "outline",
 	lineWidth: number,
 	lineColor: Color,
 }
@@ -2511,13 +2513,11 @@ type UniformValue =
 type Uniform = Record<string, UniformValue>;
 
 interface ShaderComp extends Comp {
-//  	id: "shader",
 	uniform: Uniform,
 	shader: string,
 }
 
 interface BodyComp extends Comp {
-//  	id: "body",
 	/**
 	 * If should collide with other solid objects.
 	 */
@@ -2595,7 +2595,6 @@ interface Timer {
 }
 
 interface TimerComp extends Comp {
-//  	id: "timer",
 	/**
 	 * Run the callback after n seconds.
 	 */
@@ -2603,7 +2602,6 @@ interface TimerComp extends Comp {
 }
 
 interface SolidComp extends Comp {
-//  	id: "solid",
 	/**
 	 * If should stop other solid objects from moving through.
 	 */
@@ -2611,7 +2609,6 @@ interface SolidComp extends Comp {
 }
 
 interface FixedComp extends Comp {
-//  	id: "fixed",
 	/**
 	 * If the obj is unaffected by camera
 	 */
@@ -2619,7 +2616,6 @@ interface FixedComp extends Comp {
 }
 
 interface StayComp extends Comp {
-//  	id: "stay",
 	/**
 	 * If the obj should not be destroyed on scene switch.
 	 */
@@ -2627,7 +2623,6 @@ interface StayComp extends Comp {
 }
 
 interface HealthComp extends Comp {
-//  	id: "health",
 	/**
 	 * Decrease HP by n (defaults to 1).
 	 */
@@ -2647,7 +2642,6 @@ interface HealthComp extends Comp {
 }
 
 interface LifespanComp extends Comp {
-//  	id: "lifespan",
 }
 
 interface LifespanCompConf {